Extract out linecaster from groundchecker
